swapi/requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==1.11
django-cors-headers==2.4.0
djangorestframework==3.9.0
idna==2.8
pkg-resources==0.0.0
pytz==2018.7
requests==2.21.0
urllib3==1.24.1
